- Updated test cases - Fixed capture of error body in Angular
