+++ src/templates/core/angular/request.hbs
@@ -2,7 +2,7 @@
 
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import type { HttpResponse, HttpErrorResponse } from '@angular/common/http';
-import { catchError, forkJoin, map, mergeMap, of, throwError } from 'rxjs';
+import { catchError, forkJoin, map, switchMap, of, throwError } from 'rxjs';
 import type { Observable } from 'rxjs';
 
 import { ApiError } from './ApiError';
@@ -72,7 +72,7 @@ export const request = <T>(config: OpenAPIConfig, http: HttpClient, options: Api
 	const body = getRequestBody(options);
 
 	return getHeaders(config, options).pipe(
-		mergeMap(headers => {
+		switchMap(headers => {
 			return sendRequest<T>(config, options, http, url, formData, body, headers);
 		}),
 		map(response => {
@@ -88,14 +88,14 @@ export const request = <T>(config: OpenAPIConfig, http: HttpClient, options: Api
 		}),
 		catchError((error: HttpErrorResponse) => {
 			if (!error.status) {
-				return throwError(() => error);
+				return throwError(error);
 			}
 			return of({
 				url,
 				ok: error.ok,
 				status: error.status,
 				statusText: error.statusText,
-				body: error.statusText,
+				body: error.error ?? error.statusText,
 			} as ApiResult);
 		}),
 		map(result => {
@@ -103,7 +103,7 @@ export const request = <T>(config: OpenAPIConfig, http: HttpClient, options: Api
 			return result.body as T;
 		}),
 		catchError((error: ApiError) => {
-			return throwError(() => error);
+			return throwError(error);
 		}),
 	);
 };
