+++ test/e2e/client.axios.spec.ts
@@ -3,7 +3,7 @@ import { compileWithTypescript } from './scripts/compileWithTypescript';
 import { generateClient } from './scripts/generateClient';
 import server from './scripts/server';
 
-describe('v3.node', () => {
+describe('client.node', () => {
     beforeAll(async () => {
         cleanup('client/axios');
         await generateClient('client/axios', 'v3', 'axios', false, false, 'AppClient');
@@ -82,5 +82,69 @@ describe('v3.node', () => {
             error = (e as Error).message;
         }
         expect(error).toContain('Request aborted');
+    });
+
+    it('should throw known error (500)', async () => {
+        let error;
+        try {
+            const { AppClient } = require('./generated/client/axios/index.js');
+            const client = new AppClient();
+            await client.error.testErrorCode(500);
+        } catch (e) {
+            const err = e as any;
+            error = JSON.stringify({
+                name: err.name,
+                message: err.message,
+                url: err.url,
+                status: err.status,
+                statusText: err.statusText,
+                body: err.body,
+            });
+        }
+        expect(error).toBe(
+            JSON.stringify({
+                name: 'ApiError',
+                message: 'Custom message: Internal Server Error',
+                url: 'http://localhost:3000/base/api/v1.0/error?status=500',
+                status: 500,
+                statusText: 'Internal Server Error',
+                body: {
+                    status: 500,
+                    message: 'hello world',
+                },
+            })
+        );
+    });
+
+    it('should throw unknown error (409)', async () => {
+        let error;
+        try {
+            const { AppClient } = require('./generated/client/axios/index.js');
+            const client = new AppClient();
+            await client.error.testErrorCode(409);
+        } catch (e) {
+            const err = e as any;
+            error = JSON.stringify({
+                name: err.name,
+                message: err.message,
+                url: err.url,
+                status: err.status,
+                statusText: err.statusText,
+                body: err.body,
+            });
+        }
+        expect(error).toBe(
+            JSON.stringify({
+                name: 'ApiError',
+                message: 'Generic Error',
+                url: 'http://localhost:3000/base/api/v1.0/error?status=409',
+                status: 409,
+                statusText: 'Conflict',
+                body: {
+                    status: 409,
+                    message: 'hello world',
+                },
+            })
+        );
     });
 });
