+++ test/e2e/client.babel.spec.ts
@@ -5,7 +5,7 @@ import { copyAsset } from './scripts/copyAsset';
 import { generateClient } from './scripts/generateClient';
 import server from './scripts/server';
 
-describe('v3.babel', () => {
+describe('client.babel', () => {
     beforeAll(async () => {
         cleanup('client/babel');
         await generateClient('client/babel', 'v3', 'fetch', true, true, 'AppClient');
@@ -53,13 +53,123 @@ describe('v3.babel', () => {
             const { AppClient } = (window as any).api;
             const client = new AppClient();
             return await client.complex.complexTypes({
-                first: {
-                    second: {
-                        third: 'Hello World!',
+                parameterObject: {
+                    first: {
+                        second: {
+                            third: 'Hello World!',
+                        },
                     },
                 },
             });
         });
         expect(result).toBeDefined();
+    });
+
+    it('support form data', async () => {
+        const result = await browser.evaluate(async () => {
+            const { AppClient } = (window as any).api;
+            const client = new AppClient();
+            return await client.parameters.callWithParameters({
+                parameterHeader: 'valueHeader',
+                parameterQuery: 'valueQuery',
+                parameterForm: 'valueForm',
+                parameterCookie: 'valueCookie',
+                parameterPath: 'valuePath',
+                requestBody: {
+                    prop: 'valueBody',
+                },
+            });
+        });
+        expect(result).toBeDefined();
+    });
+
+    it('can abort the request', async () => {
+        let error;
+        try {
+            await browser.evaluate(async () => {
+                const { AppClient } = (window as any).api;
+                const client = new AppClient();
+                const promise = client.simple.getCallWithoutParametersAndResponse();
+                setTimeout(() => {
+                    promise.cancel();
+                }, 10);
+                await promise;
+            });
+        } catch (e) {
+            error = (e as Error).message;
+        }
+        expect(error).toContain('CancelError: Request aborted');
+    });
+
+    it('should throw known error (500)', async () => {
+        const error = await browser.evaluate(async () => {
+            try {
+                const { AppClient } = (window as any).api;
+                const client = new AppClient();
+                await client.error.testErrorCode({
+                    status: 500,
+                });
+            } catch (e) {
+                const error = e as any;
+                return JSON.stringify({
+                    name: error.name,
+                    message: error.message,
+                    url: error.url,
+                    status: error.status,
+                    statusText: error.statusText,
+                    body: error.body,
+                });
+            }
+            return;
+        });
+        expect(error).toBe(
+            JSON.stringify({
+                name: 'ApiError',
+                message: 'Custom message: Internal Server Error',
+                url: 'http://localhost:3000/base/api/v1.0/error?status=500',
+                status: 500,
+                statusText: 'Internal Server Error',
+                body: {
+                    status: 500,
+                    message: 'hello world',
+                },
+            })
+        );
+    });
+
+    it('should throw unknown error (409)', async () => {
+        const error = await browser.evaluate(async () => {
+            try {
+                const { AppClient } = (window as any).api;
+                const client = new AppClient();
+                await client.error.testErrorCode({
+                    status: 409,
+                });
+            } catch (e) {
+                const error = e as any;
+                return JSON.stringify({
+                    name: error.name,
+                    message: error.message,
+                    url: error.url,
+                    status: error.status,
+                    statusText: error.statusText,
+                    body: error.body,
+                });
+            }
+            return;
+        });
+        expect(error).toBe(
+            JSON.stringify({
+                name: 'ApiError',
+                message: 'Generic Error',
+                url: 'http://localhost:3000/base/api/v1.0/error?status=409',
+                status: 409,
+                statusText: 'Conflict',
+                body: {
+                    status: 409,
+                    message: 'hello world',
+                },
+            })
+        );
     });
 });
