+++ test/e2e/client.fetch.spec.ts
@@ -5,7 +5,7 @@ import { copyAsset } from './scripts/copyAsset';
 import { generateClient } from './scripts/generateClient';
 import server from './scripts/server';
 
-describe('v3.fetch', () => {
+describe('client.fetch', () => {
     beforeAll(async () => {
         cleanup('client/fetch');
         await generateClient('client/fetch', 'v3', 'fetch', false, false, 'AppClient');
@@ -126,7 +126,10 @@ describe('v3.fetch', () => {
                 url: 'http://localhost:3000/base/api/v1.0/error?status=500',
                 status: 500,
                 statusText: 'Internal Server Error',
-                body: 'Internal Server Error',
+                body: {
+                    status: 500,
+                    message: 'hello world',
+                },
             })
         );
     });
@@ -157,7 +160,10 @@ describe('v3.fetch', () => {
                 url: 'http://localhost:3000/base/api/v1.0/error?status=409',
                 status: 409,
                 statusText: 'Conflict',
-                body: 'Conflict',
+                body: {
+                    status: 409,
+                    message: 'hello world',
+                },
             })
         );
     });
