+++ test/e2e/client.xhr.spec.ts
@@ -5,7 +5,7 @@ import { copyAsset } from './scripts/copyAsset';
 import { generateClient } from './scripts/generateClient';
 import server from './scripts/server';
 
-describe('v3.xhr', () => {
+describe('client.xhr', () => {
     beforeAll(async () => {
         cleanup('client/xhr');
         await generateClient('client/xhr', 'v3', 'xhr', false, false, 'AppClient');
@@ -63,6 +63,24 @@ describe('v3.xhr', () => {
         expect(result).toBeDefined();
     });
 
+    it('support form data', async () => {
+        const result = await browser.evaluate(async () => {
+            const { AppClient } = (window as any).api;
+            const client = new AppClient();
+            return await client.parameters.callWithParameters(
+                'valueHeader',
+                'valueQuery',
+                'valueForm',
+                'valueCookie',
+                'valuePath',
+                {
+                    prop: 'valueBody',
+                }
+            );
+        });
+        expect(result).toBeDefined();
+    });
+
     it('can abort the request', async () => {
         let error;
         try {
@@ -107,7 +125,10 @@ describe('v3.xhr', () => {
                 url: 'http://localhost:3000/base/api/v1.0/error?status=500',
                 status: 500,
                 statusText: 'Internal Server Error',
-                body: 'Internal Server Error',
+                body: {
+                    status: 500,
+                    message: 'hello world',
+                },
             })
         );
     });
@@ -138,7 +159,10 @@ describe('v3.xhr', () => {
                 url: 'http://localhost:3000/base/api/v1.0/error?status=409',
                 status: 409,
                 statusText: 'Conflict',
-                body: 'Conflict',
+                body: {
+                    status: 409,
+                    message: 'hello world',
+                },
             })
         );
     });
