+++ test/e2e/v2.fetch.spec.ts
@@ -44,82 +44,4 @@ describe('v2.fetch', () => {
         });
         expect(result).toBeDefined();
     });
-
-    it('can abort the request', async () => {
-        let error;
-        try {
-            await browser.evaluate(async () => {
-                const { SimpleService } = (window as any).api;
-                const promise = SimpleService.getCallWithoutParametersAndResponse();
-                setTimeout(() => {
-                    promise.cancel();
-                }, 10);
-                await promise;
-            });
-        } catch (e) {
-            error = (e as Error).message;
-        }
-        expect(error).toContain('CancelError: Request aborted');
-    });
-
-    it('should throw known error (500)', async () => {
-        const error = await browser.evaluate(async () => {
-            try {
-                const { ErrorService } = (window as any).api;
-                await ErrorService.testErrorCode(500);
-            } catch (e) {
-                const error = e as any;
-                return JSON.stringify({
-                    name: error.name,
-                    message: error.message,
-                    url: error.url,
-                    status: error.status,
-                    statusText: error.statusText,
-                    body: error.body,
-                });
-            }
-            return;
-        });
-
-        expect(error).toBe(
-            JSON.stringify({
-                name: 'ApiError',
-                message: 'Custom message: Internal Server Error',
-                url: 'http://localhost:3000/base/api/v1.0/error?status=500',
-                status: 500,
-                statusText: 'Internal Server Error',
-                body: 'Internal Server Error',
-            })
-        );
-    });
-
-    it('should throw unknown error (409)', async () => {
-        const error = await browser.evaluate(async () => {
-            try {
-                const { ErrorService } = (window as any).api;
-                await ErrorService.testErrorCode(409);
-            } catch (e) {
-                const error = e as any;
-                return JSON.stringify({
-                    name: error.name,
-                    message: error.message,
-                    url: error.url,
-                    status: error.status,
-                    statusText: error.statusText,
-                    body: error.body,
-                });
-            }
-            return;
-        });
-        expect(error).toBe(
-            JSON.stringify({
-                name: 'ApiError',
-                message: 'Generic Error',
-                url: 'http://localhost:3000/base/api/v1.0/error?status=409',
-                status: 409,
-                statusText: 'Conflict',
-                body: 'Conflict',
-            })
-        );
-    });
 });
