+++ test/e2e/v2.node.spec.ts
@@ -50,60 +50,4 @@ describe('v2.node', () => {
         }
         expect(error).toContain('Request aborted');
     });
-
-    it('should throw known error (500)', async () => {
-        let error;
-        try {
-            const { ErrorService } = require('./generated/v2/node/index.js');
-            await ErrorService.testErrorCode(500);
-        } catch (e) {
-            const err = e as any;
-            error = JSON.stringify({
-                name: err.name,
-                message: err.message,
-                url: err.url,
-                status: err.status,
-                statusText: err.statusText,
-                body: err.body,
-            });
-        }
-        expect(error).toBe(
-            JSON.stringify({
-                name: 'ApiError',
-                message: 'Custom message: Internal Server Error',
-                url: 'http://localhost:3000/base/api/v1.0/error?status=500',
-                status: 500,
-                statusText: 'Internal Server Error',
-                body: 'Internal Server Error',
-            })
-        );
-    });
-
-    it('should throw unknown error (409)', async () => {
-        let error;
-        try {
-            const { ErrorService } = require('./generated/v2/node/index.js');
-            await ErrorService.testErrorCode(409);
-        } catch (e) {
-            const err = e as any;
-            error = JSON.stringify({
-                name: err.name,
-                message: err.message,
-                url: err.url,
-                status: err.status,
-                statusText: err.statusText,
-                body: err.body,
-            });
-        }
-        expect(error).toBe(
-            JSON.stringify({
-                name: 'ApiError',
-                message: 'Generic Error',
-                url: 'http://localhost:3000/base/api/v1.0/error?status=409',
-                status: 409,
-                statusText: 'Conflict',
-                body: 'Conflict',
-            })
-        );
-    });
 });
