+++ test/e2e/v3.axios.spec.ts
@@ -3,7 +3,7 @@ import { compileWithTypescript } from './scripts/compileWithTypescript';
 import { generateClient } from './scripts/generateClient';
 import server from './scripts/server';
 
-describe('v3.node', () => {
+describe('v3.axios', () => {
     beforeAll(async () => {
         cleanup('v3/axios');
         await generateClient('v3/axios', 'v3', 'axios');
@@ -75,5 +75,67 @@ describe('v3.node', () => {
             error = (e as Error).message;
         }
         expect(error).toContain('Request aborted');
+    });
+
+    it('should throw known error (500)', async () => {
+        let error;
+        try {
+            const { ErrorService } = require('./generated/v3/axios/index.js');
+            await ErrorService.testErrorCode(500);
+        } catch (e) {
+            const err = e as any;
+            error = JSON.stringify({
+                name: err.name,
+                message: err.message,
+                url: err.url,
+                status: err.status,
+                statusText: err.statusText,
+                body: err.body,
+            });
+        }
+        expect(error).toBe(
+            JSON.stringify({
+                name: 'ApiError',
+                message: 'Custom message: Internal Server Error',
+                url: 'http://localhost:3000/base/api/v1.0/error?status=500',
+                status: 500,
+                statusText: 'Internal Server Error',
+                body: {
+                    status: 500,
+                    message: 'hello world',
+                },
+            })
+        );
+    });
+
+    it('should throw unknown error (409)', async () => {
+        let error;
+        try {
+            const { ErrorService } = require('./generated/v3/axios/index.js');
+            await ErrorService.testErrorCode(409);
+        } catch (e) {
+            const err = e as any;
+            error = JSON.stringify({
+                name: err.name,
+                message: err.message,
+                url: err.url,
+                status: err.status,
+                statusText: err.statusText,
+                body: err.body,
+            });
+        }
+        expect(error).toBe(
+            JSON.stringify({
+                name: 'ApiError',
+                message: 'Generic Error',
+                url: 'http://localhost:3000/base/api/v1.0/error?status=409',
+                status: 409,
+                statusText: 'Conflict',
+                body: {
+                    status: 409,
+                    message: 'hello world',
+                },
+            })
+        );
     });
 });
