+++ test/e2e/v3.babel.spec.ts
@@ -48,13 +48,120 @@ describe('v3.babel', () => {
         const result = await browser.evaluate(async () => {
             const { ComplexService } = (window as any).api;
             return await ComplexService.complexTypes({
-                first: {
-                    second: {
-                        third: 'Hello World!',
+                parameterObject: {
+                    first: {
+                        second: {
+                            third: 'Hello World!',
+                        },
                     },
                 },
             });
         });
         expect(result).toBeDefined();
+    });
+
+    it('support form data', async () => {
+        const result = await browser.evaluate(async () => {
+            const { ParametersService } = (window as any).api;
+            return await ParametersService.callWithParameters({
+                parameterHeader: 'valueHeader',
+                parameterQuery: 'valueQuery',
+                parameterForm: 'valueForm',
+                parameterCookie: 'valueCookie',
+                parameterPath: 'valuePath',
+                requestBody: {
+                    prop: 'valueBody',
+                },
+            });
+        });
+        expect(result).toBeDefined();
+    });
+
+    it('can abort the request', async () => {
+        let error;
+        try {
+            await browser.evaluate(async () => {
+                const { SimpleService } = (window as any).api;
+                const promise = SimpleService.getCallWithoutParametersAndResponse();
+                setTimeout(() => {
+                    promise.cancel();
+                }, 10);
+                await promise;
+            });
+        } catch (e) {
+            error = (e as Error).message;
+        }
+        expect(error).toContain('CancelError: Request aborted');
+    });
+
+    it('should throw known error (500)', async () => {
+        const error = await browser.evaluate(async () => {
+            try {
+                const { ErrorService } = (window as any).api;
+                await ErrorService.testErrorCode({
+                    status: 500,
+                });
+            } catch (e) {
+                const error = e as any;
+                return JSON.stringify({
+                    name: error.name,
+                    message: error.message,
+                    url: error.url,
+                    status: error.status,
+                    statusText: error.statusText,
+                    body: error.body,
+                });
+            }
+            return;
+        });
+
+        expect(error).toBe(
+            JSON.stringify({
+                name: 'ApiError',
+                message: 'Custom message: Internal Server Error',
+                url: 'http://localhost:3000/base/api/v1.0/error?status=500',
+                status: 500,
+                statusText: 'Internal Server Error',
+                body: {
+                    status: 500,
+                    message: 'hello world',
+                },
+            })
+        );
+    });
+
+    it('should throw unknown error (409)', async () => {
+        const error = await browser.evaluate(async () => {
+            try {
+                const { ErrorService } = (window as any).api;
+                await ErrorService.testErrorCode({
+                    status: 409,
+                });
+            } catch (e) {
+                const error = e as any;
+                return JSON.stringify({
+                    name: error.name,
+                    message: error.message,
+                    url: error.url,
+                    status: error.status,
+                    statusText: error.statusText,
+                    body: error.body,
+                });
+            }
+            return;
+        });
+        expect(error).toBe(
+            JSON.stringify({
+                name: 'ApiError',
+                message: 'Generic Error',
+                url: 'http://localhost:3000/base/api/v1.0/error?status=409',
+                status: 409,
+                statusText: 'Conflict',
+                body: {
+                    status: 409,
+                    message: 'hello world',
+                },
+            })
+        );
     });
 });
