+++ test/e2e/v3.fetch.spec.ts
@@ -118,7 +118,10 @@ describe('v3.fetch', () => {
                 url: 'http://localhost:3000/base/api/v1.0/error?status=500',
                 status: 500,
                 statusText: 'Internal Server Error',
-                body: 'Internal Server Error',
+                body: {
+                    status: 500,
+                    message: 'hello world',
+                },
             })
         );
     });
@@ -148,7 +151,10 @@ describe('v3.fetch', () => {
                 url: 'http://localhost:3000/base/api/v1.0/error?status=409',
                 status: 409,
                 statusText: 'Conflict',
-                body: 'Conflict',
+                body: {
+                    status: 409,
+                    message: 'hello world',
+                },
             })
         );
     });
